requirements.txt
tortoise-orm==0.16.13
aiogram==2.9.2
python-dotenv==0.14.0
Django==3.0.8
psycopg2==2.8.5
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.1.0
asyncpg==0.20.1
Pillow==7.1.2
django-ckeditor==5.9.0











